requirements.txt
python-telegram-bot==21.6
psycopg2==2.9.10
openai==1.64.0